AIX client handler: Cleanup unsigned long/signed long stuff in memory calculations.
git-svn-id: svn://svn.code.sf.net/p/xymon/code@6727 44351d6e-118b-4698-b696-ce33095ecaa4

+++ trunk/xymond/client/aix.c
@@ -74,17 +74,17 @@ void handle_aix_client(char *hostname, char *clienttype, enum ostype_t os,
 		long memphystotal = 0, memphysfree = 0, memswaptotal = 0, memswappct = 0;
 		char *p;
 
-		if (strncmp(realmemstr, "realmem ", 8) == 0) memphystotal = atoi(realmemstr+8) / 1024;
-		if (sscanf(freememstr, "%*d %*d %*d %ld", &memphysfree) == 1) memphysfree /= 256;
+		if (strncmp(realmemstr, "realmem ", 8) == 0) memphystotal = atol(realmemstr+8) / 1024L;
+		if (sscanf(freememstr, "%*d %*d %*d %ld", &memphysfree) == 1) memphysfree /= 256L;
 
 		p = strchr(swapmemstr, '\n'); if (p) p++;
 		if (p && (sscanf(p, " %ldMB %ld%%", &memswaptotal, &memswappct) != 2)) {
-			memswaptotal = memswappct = -1;
+			memswaptotal = memswappct = -1L;
 		}
 
 		unix_memory_report(hostname, clienttype, os, hinfo, fromline, timestr,
-				memphystotal, (memphystotal - memphysfree), -1,
-				memswaptotal, ((memswaptotal * memswappct) / 100));
+				memphystotal, (memphystotal - memphysfree), -1L,
+				memswaptotal, ((memswaptotal * memswappct) / 100L));
 	}
 
 	splitmsg_done();
